pyproject.toml
[project]
name = "semetra"
version = "0.1.0"
description = "Dein Studium. Automatisch organisiert. – Studienplan automatisch generiert aus deiner Fachhochschule."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PySide6>=6.6",
  "pdfplumber>=0.11",
  "openpyxl>=3.1",
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "python-dotenv>=1.0",
]

[project.scripts]
semetra = "semetra.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
